refactored pancake AST and parsing
