+++ pancake2viper/src/pancake/mod.rs
@@ -1,7 +1,9 @@
 mod expression;
+mod parser;
 mod shape;
 mod statement;
 mod top;
+mod utils;
 
 pub use expression::*;
 pub use shape::Shape;
