+++ pancake2viper/src/pancake/statement.rs
@@ -1,6 +1,4 @@
-use super::{parse_exp, parse_exp_list, Expr};
-use crate::parser::SExpr::{self, *};
-use anyhow::anyhow;
+use super::Expr;
 
 #[derive(Debug, Clone)]
 pub enum Stmt {
@@ -149,220 +147,4 @@ pub struct Return {
 #[derive(Debug, Clone)]
 pub struct Annotation {
     pub line: String,
-}
-
-pub fn parse_stmt_symbol(symbol: &str) -> anyhow::Result<Stmt> {
-    match symbol {
-        "break" => Ok(Stmt::Break),
-        "continue" => Ok(Stmt::Continue),
-        "skip" => Ok(Stmt::Skip),
-        "tick" => Ok(Stmt::Tick),
-        x => Err(anyhow!("Failed to parse stmt symbol: {}", x)),
-    }
-}
-
-pub fn parse_stmt(s: Vec<&SExpr>) -> anyhow::Result<Stmt> {
-    match &s[..] {
-        [Symbol(op), SString(at), SString(annot)] if op == "annot" && at == "@" => {
-            Ok(Stmt::Annotation(Annotation {
-                line: annot.to_owned(),
-            }))
-        }
-        [Symbol(op)] => parse_stmt_symbol(op),
-        // Variable declaration
-        [Symbol(op), List(decl), List(rem)] if op == "dec" => parse_dec(
-            decl.iter().collect::<Vec<_>>(),
-            Some(rem.iter().collect::<Vec<_>>()),
-        ),
-        [Symbol(op), List(decl)] if op == "dec" => parse_dec(decl.iter().collect::<Vec<_>>(), None),
-        [Symbol(var), Symbol(eq), List(exp)] if eq == ":=" => Ok(Stmt::Assign(Assign {
-            lhs: var.clone(),
-            rhs: parse_exp(exp)?,
-        })),
-        [Symbol(op), List(addr), Symbol(eq), List(exp)] if op == "mem" && eq == ":=" => {
-            Ok(Stmt::Store(Store {
-                address: parse_exp(addr)?,
-                value: parse_exp(exp)?,
-            }))
-        }
-        [Symbol(op), List(addr), Symbol(eq), Symbol(byte), List(exp)]
-            if op == "mem" && eq == ":=" && byte == "byte" =>
-        {
-            Ok(Stmt::StoreBits(StoreBits {
-                address: parse_exp(addr)?,
-                value: parse_exp(exp)?,
-                size: MemOpBytes::Byte,
-            }))
-        }
-
-        // Shared stores
-        [Symbol(op), Symbol(size), List(addr), List(exp)]
-            if op == "shared_mem_store" && size == "word" =>
-        {
-            Ok(Stmt::SharedStore(SharedStore {
-                address: parse_exp(addr)?,
-                value: parse_exp(exp)?,
-            }))
-        }
-
-        [Symbol(op), Symbol(size), List(addr), List(exp)]
-            if op == "shared_mem_store" && size == "byte" =>
-        {
-            Ok(Stmt::SharedStoreBits(SharedStoreBits {
-                address: parse_exp(addr)?,
-                value: parse_exp(exp)?,
-                size: MemOpBytes::Byte,
-            }))
-        }
-
-        [Symbol(op), Symbol(size), List(addr), List(exp)]
-            if op == "shared_mem_store" && size == "halfword" =>
-        {
-            Ok(Stmt::SharedStoreBits(SharedStoreBits {
-                address: parse_exp(addr)?,
-                value: parse_exp(exp)?,
-                size: MemOpBytes::Byte,
-            }))
-        }
-
-        // Shared loads
-        [Symbol(op), Symbol(size), Symbol(dst), List(exp)]
-            if op == "shared_mem_load" && size == "word" =>
-        {
-            Ok(Stmt::SharedLoad(SharedLoad {
-                address: parse_exp(exp)?,
-                dst: Expr::Var(dst.into()),
-            }))
-        }
-
-        [Symbol(op), Symbol(size), Symbol(dst), List(exp)]
-            if op == "shared_mem_load" && size == "byte" =>
-        {
-            Ok(Stmt::SharedLoadBits(SharedLoadBits {
-                address: parse_exp(exp)?,
-                dst: Expr::Var(dst.into()),
-                size: MemOpBytes::Byte,
-            }))
-        }
-
-        [Symbol(op), Symbol(size), Symbol(dst), List(exp)]
-            if op == "shared_mem_load" && size == "halfword" =>
-        {
-            Ok(Stmt::SharedLoadBits(SharedLoadBits {
-                address: parse_exp(exp)?,
-                dst: Expr::Var(dst.into()),
-                size: MemOpBytes::HalfWord,
-            }))
-        }
-
-        [Symbol(op), stmts @ ..] if op == "seq" => parse_seq(stmts),
-
-        // if
-        [Symbol(op), List(cond), List(b1), List(b2)] if op == "if" => Ok(Stmt::If(If {
-            cond: parse_exp(cond)?,
-            if_branch: Box::new(parse_stmt(b1.iter().collect::<Vec<_>>())?),
-            else_branch: Box::new(parse_stmt(b2.iter().collect::<Vec<_>>())?),
-        })),
-        [Symbol(op), List(cond), List(b1), Symbol(b2)] if op == "if" => Ok(Stmt::If(If {
-            cond: parse_exp(cond)?,
-            if_branch: Box::new(parse_stmt(b1.iter().collect::<Vec<_>>())?),
-            else_branch: Box::new(parse_stmt_symbol(b2)?),
-        })),
-        [Symbol(op), List(cond), Symbol(b1), List(b2)] if op == "if" => Ok(Stmt::If(If {
-            cond: parse_exp(cond)?,
-            if_branch: Box::new(parse_stmt_symbol(b1)?),
-            else_branch: Box::new(parse_stmt(b2.iter().collect::<Vec<_>>())?),
-        })),
-        [Symbol(op), List(cond), Symbol(b1), Symbol(b2)] if op == "if" => Ok(Stmt::If(If {
-            cond: parse_exp(cond)?,
-            if_branch: Box::new(parse_stmt_symbol(b1)?),
-            else_branch: Box::new(parse_stmt_symbol(b2)?),
-        })),
-        // while
-        [Symbol(op), List(cond), List(body)] if op == "while" => Ok(Stmt::While(While {
-            cond: parse_exp(cond)?,
-            body: Box::new(parse_stmt(body.iter().collect::<Vec<_>>())?),
-        })),
-        [Symbol(op), List(cond), Symbol(body)] if op == "while" => Ok(Stmt::While(While {
-            cond: parse_exp(cond)?,
-            body: Box::new(parse_stmt_symbol(body)?),
-        })),
-        [Symbol(op), List(label), List(args), Symbol(ret)] if op == "call" => {
-            Ok(Stmt::Call(Call {
-                rettype: ret.into(),
-                fname: parse_exp(label)?,
-                args: parse_exp_list(args)?,
-            }))
-        }
-        [Symbol(op), List(label), List(args), Int(ret)] if op == "call" => Ok(Stmt::Call(Call {
-            rettype: "todo_call".into(),
-            fname: parse_exp(label)?,
-            args: parse_exp_list(args)?,
-        })),
-        [Symbol(op), List(exp)] if op == "return" => Ok(Stmt::Return(Return {
-            value: parse_exp(exp)?,
-        })),
-        [Symbol(op), Symbol(name), List(arg0), List(arg1), List(arg2), List(arg3)]
-            if op == "ext_call" =>
-        {
-            Ok(Stmt::ExtCall(ExtCall {
-                fname: name.clone(),
-                args: [
-                    parse_exp(arg0)?,
-                    parse_exp(arg1)?,
-                    parse_exp(arg2)?,
-                    parse_exp(arg3)?,
-                ],
-            }))
-        }
-        [Symbol(op), List(label), List(args)] if op == "tail_call" => {
-            Ok(Stmt::TailCall(TailCall {
-                fname: parse_exp(label)?,
-                args: parse_exp_list(args)?,
-            }))
-        }
-        [List(stmt)] => parse_stmt(stmt.iter().collect::<Vec<_>>()),
-        x => panic!("Could not parse stmt: {:?}", x),
-    }
-}
-
-fn parse_dec(decl: Vec<&SExpr>, scope: Option<Vec<&SExpr>>) -> anyhow::Result<Stmt> {
-    let scope = match scope {
-        Some(stmts) => parse_stmt(stmts)?,
-        None => Stmt::Skip,
-    };
-    match &decl[..] {
-        [Symbol(var), Symbol(eq), List(exp)] if eq == ":=" => Ok(Stmt::Declaration(Declaration {
-            lhs: var.clone(),
-            rhs: parse_exp(exp)?,
-            scope: Box::new(scope),
-        })),
-        _ => Err(anyhow!("Not a valid declaration")),
-    }
-}
-
-fn parse_seq(s: &[&SExpr]) -> anyhow::Result<Stmt> {
-    // TODO: handle annotations or positional information correctly i don't remember
-    match s {
-        // this might be positional sequence
-        [List(pos), pstmt] => {
-            match &pos[..] {
-                [Symbol(loc), st @ ..] if loc == "annot" => {
-                    return parse_stmt(vec![pstmt]);
-                }
-                // [Symbol(loc), Symbol(start), Symbol(stop)] => panic!("{}->{}", start, stop),
-                _ => (),
-            }
-        }
-        _ => (),
-    }
-    let mut stmts = vec![];
-    for stmt in s {
-        match stmt {
-            List(stmt) => stmts.push(parse_stmt(stmt.iter().collect())?),
-            Symbol(stmt) => stmts.push(parse_stmt_symbol(stmt)?),
-            _ => return Err(anyhow!("Error whilst parsing sequence")),
-        }
-    }
-    Ok(Stmt::Seq(Seq { stmts }))
 }
